Updated setup dependencies
