+++ dshidmini.sln
@@ -31,6 +31,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "scpdlltester", "scpdlltester\scpdlltester.vcxproj", "{FEB89FC0-BF9B-43F3-8467-B3496D61B76E}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Setup", "Setup", "{58E023F1-01BB-4D75-90A5-2E6049F94048}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DsHidMini.Installer", "setup\DsHidMini.Installer.csproj", "{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug_OTEL|Any CPU = Debug_OTEL|Any CPU
@@ -191,6 +195,38 @@ Global
 		{FEB89FC0-BF9B-43F3-8467-B3496D61B76E}.Release|x64.Build.0 = Release|x64
 		{FEB89FC0-BF9B-43F3-8467-B3496D61B76E}.Release|x86.ActiveCfg = Release|Win32
 		{FEB89FC0-BF9B-43F3-8467-B3496D61B76E}.Release|x86.Build.0 = Release|Win32
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|Any CPU.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|Any CPU.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|ARM64.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|ARM64.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|x64.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|x64.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|x86.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug_OTEL|x86.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|ARM64.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|ARM64.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|x64.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Debug|x86.Build.0 = Debug|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|Any CPU.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|Any CPU.Build.0 = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|ARM64.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|ARM64.Build.0 = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|x64.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|x64.Build.0 = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|x86.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release_OTEL|x86.Build.0 = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|ARM64.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|ARM64.Build.0 = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|x64.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|x64.Build.0 = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|x86.ActiveCfg = Release|Any CPU
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -200,6 +236,7 @@ Global
 		{8C041739-B65C-4BE3-B068-B13ED698C72E} = {CE492389-7FB3-4DC4-9AFF-B7A04F70F891}
 		{DC58FE95-938F-4BFB-B434-C043C62E9EF5} = {D5FFDFFE-55A3-4AF3-92A6-13F28598A5EE}
 		{FEB89FC0-BF9B-43F3-8467-B3496D61B76E} = {D5FFDFFE-55A3-4AF3-92A6-13F28598A5EE}
+		{7D07C49F-A5A8-44AD-83B5-1E5B1F3080AA} = {58E023F1-01BB-4D75-90A5-2E6049F94048}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {5A4A580B-8A5D-4E6A-A75D-ADFDD46F6F37}
