+++ setup/Util/StringExtensions.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Nefarius.DsHidMini.Setup.Util;
+
+public static class StringExtensions
+{
+    public static bool Contains(this string source, string toCheck, StringComparison comp)
+    {
+        return source?.IndexOf(toCheck, comp) >= 0;
+    }
+}
